[src,benchmark,doc] chore: float64 cast in IoU/F1, detection OOD task lists, parallelism doc, README
- Fix: cast cal_IoU and cal_F1 inputs to float64 in parse_utils to
  prevent precision issues with integer-typed array inputs
- Add tasks_list/tasks_MedVision-detect-planeOOD-CoT.json
- Add tasks_list/tasks_MedVision-detect-taskOOD-CoT.json
- Add docs/Model-Parallelism-Summary.md documenting DP vs TP strategy
  and DDP wrapper usage for each benchmark model
- Add vllm and verl to README acknowledgements

+++ src/medvision_bm/utils/parse_utils.py
@@ -76,8 +76,8 @@ def convert_numpy_to_python(obj):
 
 def cal_IoU(pred, target):
     # Ensure inputs are 1D numpy arrays with 4 numbers
-    pred = np.asarray(pred).flatten()
-    target = np.asarray(target).flatten()
+    pred = np.asarray(pred, dtype=np.float64).flatten()
+    target = np.asarray(target, dtype=np.float64).flatten()
 
     if len(pred) != 4 or len(target) != 4:
         raise ValueError(
@@ -122,8 +122,8 @@ def cal_IoU(pred, target):
 
 def cal_F1(pred, target):
     # Ensure inputs are 1D numpy arrays with 4 numbers
-    pred = np.asarray(pred).flatten()
-    target = np.asarray(target).flatten()
+    pred = np.asarray(pred, dtype=np.float64).flatten()
+    target = np.asarray(target, dtype=np.float64).flatten()
 
     if len(pred) != 4 or len(target) != 4:
         raise ValueError(
